Add files via upload
This program uses Python's turtle module to create a turtle racing game. The user is requested to guess the winner of seven turtles. A feedback is given after the race and the user is informed of the outcome of the race; the user gets to know whether he/she lost or won.